Honour the colour emphasis bits, and put AccuracyCoin in its place
The three high bits of the mask register were read and thrown away. They are
not a palette entry and no table of 64 colours can express them: they hold two
channels back so the third stands out, and with all three set the picture
simply darkens. Games use them for a screen-wide flash, for going under water,
and for the frame a hit lands on, so ignoring them was visible in ordinary
play rather than only in a test.

Because a game can change them partway down a frame, the setting belongs to the
pixel. Each entry of the finished picture now carries its palette index in the
low six bits and the emphasis in force when the beam passed in the three above
it, and the display colours it in from a table of the 64 colours under each of
the eight settings. Save states move to v13 for the wider picture; a v12 or
older one still loads, its pixels read back a byte at a time with no emphasis.

The README also stops reporting AccuracyCoin beside the standard suites. That
collection aims at one specific console and probes analog-level quirks almost
no commercial game relies on; sitting in the same table as the 55-ROM bar it
invited reading the two as one percentage. The measurement stays, with its
itemized report, under a heading that says what it is — and the stale numbers
in that section are corrected: 118 to 124 of 144, and 717 to 730 offline
checks.

Unchanged: 55/55, 16/16, 2/2 ROM suites, 124/144 AccuracyCoin and 95 input
checks; internal checks 730 to 735.

+++ tests/NesEmulator.Tests/SpriteEvaluationTests.cs
@@ -127,7 +127,7 @@ internal static class SpriteEvaluationTests
             check("sprite output: X counters run while only backgrounds are shown", nes.Ppu.FrameBuffer[11 * 256 + 8] == 0x21);
         }
         {
-            byte[] Frame(bool hideBackground)
+            ushort[] Frame(bool hideBackground)
             {
                 Nes nes = Machine();
                 nes.Ppu.WriteRegister(0x2006, 0x3F); nes.Ppu.WriteRegister(0x2006, 1);
@@ -139,7 +139,7 @@ internal static class SpriteEvaluationTests
                 At(nes.Ppu, 0, 33);
                 return nes.Ppu.FrameBuffer.AsSpan(8, 24).ToArray();
             }
-            byte[] reference = Frame(false);
+            ushort[] reference = Frame(false);
             check("background output: shifters run while only sprites are shown",
                 reference.All(b => b == 0x12) && Frame(true).SequenceEqual(reference));
         }
@@ -186,8 +186,8 @@ internal static class SpriteEvaluationTests
                     nes.Ppu.Step();
                 }
 
-                ReadOnlySpan<byte> line = nes.Ppu.FrameBuffer.AsSpan(11 * 256, 256);
-                return line.IndexOf((byte)0x21);
+                ReadOnlySpan<ushort> line = nes.Ppu.FrameBuffer.AsSpan(11 * 256, 256);
+                return line.IndexOf((ushort)0x21);
             }
 
             int plain = FirstSpritePixel(blank: false);
